rebase -i: don't invoke editor unecessarily when continuing
Add a config option rebase.continue.alwaysEdit that when set to false changes
merge based rebases to only prompt the user to edit the commit message if rebase
would have done so if there hadn't been conflicts. I don't think I've ever
wanted to reword the message after conflicts and the apply backend does not open
the editor when there are conflicts. The new config option defaults to true so
does not change the existing behavior unless it is set.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>