sequencer: fix cleanup with --signoff and -x
Before commit 356ee4659b ("sequencer: try to commit without forking 'git
commit'", 2017-11-24) when --signoff or -x were given on the command
line the commit message was cleaned up with --cleanup=space or
commit.cleanup if it is set. Unfortunately this behavior was lost I
implemented committing without forking. Fix this and add some tests to
catch any regression in the future.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>