t3404: set $EDITOR in subshell
As $EDITOR is exported setting it in one test affects all subsequent
tests. Avoid this by always setting it in a subshell. This allows us to remove
calls to set_fake_editor in 21 tests that were not using the editor to
rearrange the todo list or set a commit message. They can now safely use the
default editor. Also add a test at the end of the file to check that $EDITOR and
related variables and config settings are unchanged. Hopefully in future more
rebase test files will call test_editor_unchanged at the end.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>