rebase -r: try to cherry-pick merges
When rebase -r recreates a merge it looses any conflict resolution or
semantic fixups. If only one parent tree has changed then we can try and
preserve those by cherry-picking the existing merge commit with respect
to the changed parent. If there are conflicts when cherry-picking then
we throw the result away and fall back to a normal merge. This avoids
the potential confusion of sometimes showing conflicts from merging the
requested commit and other times showing conflicts from cherry-picking
the existing merge.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>