worktree: show git-dir with list --porcelain
If a script needs to check some state in other worktrees then it needs
to be able to get the git-dir of each worktree. Running `git rev-parse
--git-dir` in each worktree is inefficient. Also if the worktree path is
on removable media that is missing then it cannot get the git-dir by
running `git rev-parse --git-dir` from the worktree anyway.

For example I have a script that lets the user amend a commit and then
updates all the branches containing that commit. The script needs to
check that the branches to be rewritten are not already being rewritten
by a rewrite running in a another worktree. It does this by first
writing the list branches to be rewritten to $GIT_DIR/git-rewrite/state
and then reading the state files from other worktrees. To do this it
needs to know the git-dir of each worktree. Running `git rev-parse
--git-dir` from each worktree is inefficient and does not work if the
worktree is on removable media is not present.

Change this to --git-path.  Fix do_git_path() to take a worktree and reject
env-vars if !wt or is_current_worktree(wt). Update revision.c to avoid
is_current_worktree() check on the index. For hook dir (and bare in worktree.c)
if repository_format_worktree_config is set call config_with_options() with
ignore_worktree (and maybe ignore commandline?) set. Then call
git_config_from_file() on config.worktree.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>