worktree: fix bare repo detection with extensions.worktreeConfig
If there are per-worktree config files then the main worktree inherited
the value of core.bare of the current worktree. This meant the value of
wt->is_bare could be incorrect when listing the worktrees from a linked
worktree. With per-worktree config files it is possible (though not
recommended) to set core.bare to true in .git/config and false in each
linked worktree's config.worktree so we have to parse the config of each
worktree to be sure to get the correct value. We could just assume that
no linked worktree is bare but then 'git worktree list' may give
different results to 'git rev-parse --is-bare-repository' run in the
worktree.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>