path: worktree_git_path() relocate paths
543107333b ("path: worktree_git_path() should not use file relocation",
2017-06-22) stopped relocating paths returned by worktree_git_path() on the
basis that the relocation for the index was wrong when $GIT_INDEX_FILE was set
and the callers only cared about getting paths under
.git/worktrees/<id>/. However we're about to add a caller with a user specified
path so we need to be able to adjust paths that should but under
$GIT_COMMON_DIR. Fix this by only respecting $GIT_INDEX_FILE for the current
worktree. We honor $GIT_OBJECT_DIRECTORY and $GIT_GRAFTS_FILE for all worktrees
as those point to shared resources. If there is per-wortkree config for
core.hooksPath then we currently return the hooks path for the current worktree
in all cases. This will be fixed in the next commit.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>