Improve stash -p stashing
unstashed stashed   expected actual
@@ 10,…    @@ 11,…  @@10,…   @@10,…
 x         x         x        x
+a         +a       +a        x
+a         +a       +a       +a
 x         x         x       +a
+a                            x
+a
 x
When stash reverts the stashed changes in the worktree it will remove
the wrong ’+a’ due to a lack of context (even if the context was
correct it would fail I think)

unstashed stashed
@@ 10,…   @@ 10,…
 x         x
+a        +a
+b        +c
+c

stash fails to revert the worktree as the patch does not apply

The solution to both these is to create a better patch. This gives
better results that 3-way merge as it avoids conflicts

walk stashed and unstashed diffs
 - if the lines match reverse the unstashed line
 - if the stash contains a line that is not in the unstashed fallback
   to 3-way merge
 - if an addition appears in unstashed but not stashed then convert it
   to a context line
 - if a deletion appears in unstashed but not stashed then drop it

apply the modified unstashed diff to the worktree to revert the
stashed changes.

This does not deal with the case where the stashed patch is ambiguous

unstashed stashed  revert A  revert B
@@ 10,…   @@ 10,…  @@ 10,…   @@ 10,…
 x         x        x         x
+a        +a       +a        +b
+b         x       +b        +a
+a                  x         x
 x

Which ’a’ should be deleted from the worktree - we cannot know unless
the user selects lines to stash rather than editing the hunk. We can
detect this case by walking the two diffs backwards as well as
forwards and comparing the resulting patch - if they differ then abort
or go for a 3-way merge (can we fake that?)