improve reverse patch editing
Rather than editing the reverse patch, allow the user to edit the hunk
as if it is going to be applied normally and adjust it afterwards. To
do this we reverse the unedited patch and apply the edited version on
top of that. Need to keep a copy of the plain original in case the user
re-edits the edited hunk. Also need to keep the edited display and
plain versions and a plain version of the combined hunk that will be
applied.