add -p: fix a test
f4d35a6b49 ("add -p: fix counting empty context lines in edited
patches", 2018-06-11) introduced a test to check that the hunk
recounting worked if the edited patch deletes the leading space of
empty context lines. However the edited patch includes an error - the
last context line does not begin with a space. The test passed because
the hunk recounting correctly counted the empty context line and did
not count the last line so it still applied so it was actually testing
that we handled empty context lines without a leading space but it
would be clearer if the last line was correct. This error was found by
the checks on edited patches that are added in the next few commits.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>