t3701: prepare a test for better hunk checks
In an upcoming commit we're going to be checking that the pre-image of
edited hunks unambiguously matches the original so we need to have a
full compliment of trailing context lines in this test or the edited
pre-image could match the original at other offsets. The change to the
comment reflects that this patch is used in a test for pathological
context lines. It checks that the offsets of later hunks are correctly
adjusted.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>