patience diff: remove myers fallback
When the patience_diff() is called on a range of lines that does not
contain any unique context lines it falls back to calling the myers
algorithm on that region. The myers implementation calls
xdl_optimize_ctxs() but because it is called on a subset of the input it
cannot optimize the context lines as well is it does when called on the
whole file. In particular insignificant blank lines that would be
ignored as matches by the myers algorithm when it is run on the whole
file will be considered matches within a smaller region. This has the
unfortunate effect of matching blank lines as context when they are not
meaningful generating sub-optimal diffs.

Instead when there are no unique context lines within a region just find
the leading, trailing and inter-hunk context. This is more in keeping
with the patience algorithm outlined by Bram Cohen[1].

[1] https://bramcohen.livejournal.com/73318.html

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>