cmake: fix an error on non-windows platforms
The code for non-windows platforms added in cd0a852981 ("cmake: add
knob to disable vcpkg", 2021-06-06) does not work as expected as the
set command which is supposed to stop it from running vcpkg is missing
a mandatory description argument. This makes the call to set fail
which means it always tries to run vcpkg which makes cmake error out
as it is missing.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>