rebase --fork-point: improve error message for invalid upstream
Error out if upstream is an ambiguous ref, before only did this if
fork point was set. Accept ambiguous if it starts refs/. Easily
defeated with ambiguous^0

TODO memory leak

Giving --fork-point without an upstream ref results in

       fatal: No such ref 'upstream that is not a ref'

which is not as helpful as it could be. Change this to

      fatal: --fork-point requires <upstream> to be a branch

Also if <upstream> is a tag then --fork-point has no effect so add a
warning if <upstream> is not a branch ref.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>