editor: use wait_after_clean
While the editor is running we ignore SIGINT and SIGQUIT to ensure
that git does not exit before the editor. When the editor exits we
check to see if it was killed by either of those signals and if it was
we re-raise the signal. Instead we can set wait_after_clean when
running the editor and run_command() will take care of waiting for the
editor to exit before re-raising the signal. As well as being simpler
this has the advantage that we don't have to rely on the editor's exit
code to see if we have received a signal while it was running.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>