fetch: cleanup filter_prefetch_refspec()
This code caught my attention because of a false positive double free
warning from -fanalyser. There is nothing particularly wrong with the
code apart from leaking the raw refspec string when it shifts the
array down but it does not take the best advantage of our utility
functions and macros. As a bonus the new version also avoids the false
positive from -fanalyser.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>