packfile: avoid undefined pointer comparisons
check_pack_index_ptr() is designed to check if the pointer used to
access the pack file index entry points within the file. Unfortunately
if the pointer passed does not point inside the file then it is
invalid and the program is undefined from the point that it was
created. Fix this by calculating the number of entries in the 8 byte
offset table when we load the index file and comparing the offset
within the table to that instead.

Found with  -fsanitize=pointer-compare.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>