tests: add test_todo to mark known breakages
test_todo is intended as a fine grained replacement for
test_expect_failure. Rather than marking the whole test as failing
test_todo is used to mark the individual commands within a test that
are failing. This approach to writing failing tests allows us to
detect unexpected failures that are hidden by test_expect_failure.

Failing commands are reported by the test harness in the same way as
test_expect_failure so there is no change in output when migrating
from test_expect_failure to test_todo. If a command marked with
test_todo succeeds then the test will fail. This is designed to make
it easier to see when a command starts succeeding in our CI compared
to using test_expect_failure where it is easy to fix a failing test
case and not realize it.

test_todo is built upon test_expect_failure but accepts commands
starting with test_* in addition to git. An our test_* assertions use
BUG() to signal on error in their use any such error will not be
hidden by test_todo.

One limitation of test_todo is that it cannot be used in a subshell.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>