rebase -i: allow a comment after a "break" command
When adding a "break" command to a rebase todo list it can be helpful to
add a comment to remind the user what they were planning to do when
the rebase stopped. Anything following the command is now interpreted as
a comment that is printed when the rebase stops. This rules out adding
arguments to the "break" command in the future but it is hard to see
what they'd be used for.

Now that all the command expect exec accept comments the help at the
bottom of the todo list is updated.

Reported-by: Olliver Schinagl <oliver@schinagl.nl>
Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>