rebase -m: write state in one place
Most of the state files are written by write_basic_state() in
sequencer.c but historically a couple of state files were written by
different rebase--helper commands. As we now do everything in a single
process and have a fully populated struct replay_opts before writing any
state it is clearer to write all the files in the same function. There
is a small change in behavior - if we cannot create the "interactive"
file it will now die straight away rather than returning an error. As we
already die if we cannot write the other state files this should not be
a problem in practice.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>