rebase -m: inline init_basic_state() into do_interactive_rebase()
After the last commit there is not much left in the body of
init_basic_state() so move it into do_interactive_rebase(). There is a
small change to the moved code as we now check for EEXIST when creating
the state directory (which may have already been created by
create_autostash()) rather than calling is_directory() before trying to
create it.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>