sequencer: start allowing commands to take comments
Allow the user to add comments to the end of "label", "reset" and
"update-ref" commands in the todo list. The parser is permissive in what
it accepts, treating anything after the required arguments as a
comment. This matches the current behavior for "reset" which ignores
anything after the target commit. For "label" and "update-ref" this is a
change in behavior and there is a small risk that invalid arguments to
the "label" and "update-ref" commands that contain a space will now be
treated as a valid argument plus comment.

When writing the todo list it will separate the comment from the rest of
the line with "#". A fixed comment character is used to match the
comments currently accepted by the "merge" command. The next couple of
commits will add support for comments after other commands.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>