rebase -m: fix --signoff with conflicts
Committing a conflict resolution with "rebase --continue" for a rebase
started with "git rebase --signoff" fails to add the "Signed-off-by:"
trailer. This happens because we we reuse the original commit message
instead of using the commit message that has been prepared by the
sequencer. Fix this by copying .git/MERGE_MSG which contains the message
file with the "Signed-off-by:" line to .git/rebase-merge/message when we
stop for the user to resolve conflicts.

TODO: explain why MERGE_MSG might be missing or why rebase-merge/message
might exist already (fixups for the latter)

successful edit => no MERGE_MSG
failed fixup => message already exists

Note still bug when committing changes after a successful edit the
signed-off-by: will be removed - fixed by sequencer context. This series
should be reworked to add failing tests that are then fixed by
sequencer-context

Note that this introduces a change in behavior as the message file now
contains conflict comments. This is safe because commit_staged_changes()
passes an explicit cleanup flag when not editing the message and when
the message is being edited it will be cleaned up automatically. This
means user now sees the same message in editor with "rebase --continue"
as they would if they ran "git commit" themselves before continuing.

The tests are extended to check that all commits made after continuing a
rebase have a "Signed-off-by:" trailer

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>