copy: improve reporting of open errors
If copy_file() cannot open the source or destination file then return an
error code indicating which file it failed to open. This changes the
return value in those cases but all of the existing callers just
check for a non-zero return value so are not affected by this change.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>