sequencer: start removing private fields from public API
"struct replay_opts" has a number of fields that are for internal
use. While they are marked as private having them in a public struct
means that every time the internal details are changed we have to
recompile all the files that include sequencer.h even though the public
API is unchanged. This commit starts the process of removing the private
fields by adding a pointer to a "struct replay_ctx" to "struct
replay_opts" and moving the "reflog_message" to the new private struct.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>