rebase -i: ignore signals when forking subprocesses
If the user presses Ctrl-C to interrupt a program run by rebase "exec"
command then SIGINT will also be sent to the git process running the
rebase resulting in it being killed. Fortunately the consequences of
this are not severe as all the state necessary to continue the rebase is
saved to disc but it would be better to avoid killing git and instead
report that the command failed. A similar situation occurs when the
sequencer runs "git commit" to reword a commit message. If the user
generates SIGINT while editing the message then the git processes
creating the commit will ignore it but the git process running the
rebase will be killed.

Fix this by ignoring SIGINT and SIGQUIT when forking "exec" commands and
"git commit". This matches what "git commit" does when running the
user's editor and matches the behavior of the standard library's
system() function.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>