rebase -i: fix commit message if a squash fails
If a "squash" command fails because it would overwrite an untracked
file then its commit message ends up being added to the final commit
twice. This happens because the message from the commit is appended to
the squash message file before we know if it fails. Then when it is
executed successfully it is added again. If the command ends up being
skipped message is still appended.

There is an existing test to check the final message when a squash
command fails but unfortunately it only checks the final line of the
message so does not detect that the message is doubled.

TODO: Fix this properly by restoring the previous message when
rescheduling and checking for fixup_count > 1 (or just 0) without
amend_head existiting in commit_staged_changes()

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>